# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User signed up successfully. JWT sent in cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Signup successful"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - validation or conflict",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 400,
                        "message": "These categories are invalid: Cleaning, Laundry",
                        "error": "Bad Request"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 400,
                        "message": "NID already exists",
                        "error": "Bad Request"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 400,
                        "message": "CRN already exists",
                        "error": "Bad Request"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 400,
                        "message": "Email already exists",
                        "error": "Bad Request"
                      }
                    }
                  ]
                },
                "examples": {
                  "categories": {
                    "summary": "Invalid Categories",
                    "value": {
                      "statusCode": 400,
                      "message": "These categories are invalid: Cleaning, Laundry",
                      "error": "Bad Request"
                    }
                  },
                  "nid": {
                    "summary": "NID already exists",
                    "value": {
                      "statusCode": 400,
                      "message": "NID already exists",
                      "error": "Bad Request"
                    }
                  },
                  "crn": {
                    "summary": "CRN already exists",
                    "value": {
                      "statusCode": 400,
                      "message": "CRN already exists",
                      "error": "Bad Request"
                    }
                  },
                  "email": {
                    "summary": "Email already exists",
                    "value": {
                      "statusCode": 400,
                      "message": "Email already exists",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Registers a new user and returns a JWT token in a cookie",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Either \"email\" or \"crn\" must be provided, but not both.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": [
                      "email",
                      "password"
                    ],
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email"
                      },
                      "password": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "crn",
                      "password"
                    ],
                    "properties": {
                      "crn": {
                        "type": "string"
                      },
                      "password": {
                        "type": "string"
                      }
                    }
                  }
                ]
              },
              "examples": {
                "loginWithEmail": {
                  "summary": "Login using email and password",
                  "value": {
                    "email": "user@example.com",
                    "password": "securePassword123"
                  }
                },
                "loginWithCrn": {
                  "summary": "Login using CRN and password",
                  "value": {
                    "crn": "1234567890",
                    "password": "securePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login successful"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request – validation or authentication failure",
            "content": {
              "application/json": {
                "examples": {
                  "userNotFound": {
                    "summary": "The user does not exist in the system",
                    "value": {
                      "statusCode": 400,
                      "message": "User not found",
                      "error": "Bad Request"
                    }
                  },
                  "invalidCredentials": {
                    "summary": "The email or CRN exists, but the password is incorrect",
                    "value": {
                      "statusCode": 400,
                      "message": "Invalid credentials",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login user and send back JWT token as a cookie",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/verify-email": {
      "post": {
        "description": "This endpoint should be called when the user clicks the verification button which is after clicking the link in their email.<br> So this endpoint should be called from the `/verify-email` route in the frontend, remember to pass the token as a query parameter (so the link will be `/verify-email?token=`).",
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "JWT token from email link",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Email verified successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token, or already verified",
            "content": {
              "application/json": {
                "examples": {
                  "userNotFound": {
                    "summary": "The user not found or already verified",
                    "value": {
                      "statusCode": 400,
                      "message": "User not found or already verified",
                      "error": "Bad Request"
                    }
                  },
                  "invalidCredentials": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "statusCode": 400,
                      "message": "Invalid or expired verification token",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Verify user email using token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/resend-verification-email": {
      "post": {
        "description": "This endpoint allows users to request a new verification email if they missed the original one or if the token expired. <br>If the you received \"Invalid or expired verification token\" error you should call this endpoint to resend the email. ",
        "operationId": "AuthController_resendVerificationEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Email to resend verification to",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "example@email.com"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Verification email resent successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already verified",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Email already verified",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Resend verification email to user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/request-password-reset": {
      "post": {
        "description": "This endpoint sends a password reset link to the user's email address.<br><br>\n            It is used in two cases:<br>\n            - If the user clicks the \"Forgot Password?\" button.<br>\n            - If the user previously requested a reset link but the 5-minute token has expired.<br><br>\n            Note: The token will only be sent if the email exists and is verified. For security reasons, we always return the same success message, regardless of whether the user exists or is verified.",
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The email address of the user who wants to reset their password.",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "user@example.com"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reset password email sent successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Password reset email sent successfully"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to send reset password email",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "message": "Failed to send reset password email"
                  }
                }
              }
            }
          }
        },
        "summary": "Request password reset link",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "description": "This endpoint is used after the user clicks the password reset link from their email.<br><br>\n    It should be triggered from the frontend's `/reset-password` route.<br>\n    The token should be passed as a query parameter in the URL.<br><br>\n    The body must include a new password that is between 8 and 28 characters long.",
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "The JWT token sent in the reset password email. This must be passed as a query parameter.",
            "schema": {
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The new password to set for the user account.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Password reset successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token, or validation error in request body.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid or expired reset password token"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found (after token is verified).",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found"
                  }
                }
              }
            }
          }
        },
        "summary": "Reset user password using token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/switch-role": {
      "patch": {
        "operationId": "AuthController_switchUserRole",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "JWT token must be set in the cookie named \"token\". Example: token=your_jwt_token_here",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token regenerated with updated role",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Role switched successfully",
                    "newRole": "SUPPLIER"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: Token missing or invalid/expired.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "No token found in cookies",
                        "error": "Unauthorized"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "Invalid or expired token",
                        "error": "Unauthorized"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected role: GUEST should never reach this endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "message": "Unexpected role: GUEST should never reach this endpoint",
                    "error": "Internal Server Error"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "cookie": []
          },
          {
            "bearer": []
          },
          {
            "token": []
          }
        ],
        "summary": "Switch user role (BUYER ⇌ SUPPLIER)",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Silah Backend API Documentation",
    "description": "Use this documentation to explore, test, and understand the available API endpoints, their request/response structure, and any required parameters such as headers, cookies, or authentication tokens.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8,
            "maxLength": 28
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "crn": {
            "type": "string",
            "example": "1234567890"
          },
          "businessName": {
            "type": "string",
            "example": "Acme Corp"
          },
          "city": {
            "type": "string",
            "example": "Riyadh"
          },
          "nid": {
            "type": "string",
            "example": "0987654321"
          },
          "categories": {
            "example": [
              "Home & Living",
              "Technical & Repair Services"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "agreedToTerms": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "crn",
          "businessName",
          "city",
          "nid",
          "categories",
          "agreedToTerms"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "The new password to replace the old one.",
            "example": "StrongPass123",
            "minLength": 8,
            "maxLength": 28
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}